// src/app/enumerator/posone.enum.ts
export enum POSOneEnum {
  REDIRECT_BY_GUID = 'guId',
  REDIRECT_PAGE_BY_TYPE_MSG = 'type',
  NAME_MACHINE_MODERN_X = 'MODERN_X',
  NAME_MACHINE_SMART = 'SMART',
  DESCRIPTION_MACHINE_MODERN_X = 'Moderninha X',
  DESCRIPTION_MACHINE_SMART = 'Moderninha Smart',
  PRACE_MICHINE_MODERN_X = 39.9,
  PRACE_MICHINE_SMART = 49.9,
  PLAN_NAME_1 = 'BASICO',
  PLAN_NAME_2 = 'INTERMEDIARIO',
  PLAN_NAME_3 = 'PRO',
  DESCRIPTION_PLAN_NAME_1 = 'Básico',
  DESCRIPTION_PLAN_NAME_2 = 'Intermediário',
  DESCRIPTION_PLAN_NAME_3 = 'Profissional',
  PLAN_PRACE_1 = 30.90,
  PLAN_PRACE_2 = 60.90,
  PLAN_PRACE_3 = 80.90,
  DESCRIPTION_PLAN_DENEFIT_1 = 'Benefício A',
  DESCRIPTION_PLAN_DENEFIT_2 = 'Benefício B',
  DESCRIPTION_PLAN_DENEFIT_3 = 'Benefício C',
}